+++ src/main/java/net/rogues/item/armor/Armors.java
@@ -52,6 +52,12 @@ public class Armors {
             10,
             RogueSounds.ROGUE_ARMOR_EQUIP.entry(), () -> { return Ingredient.ofItems(Items.RABBIT_HIDE); });
 
+    public static RegistryEntry<ArmorMaterial> material_rogue_t3 = material(
+            "netherite_assassin_armor",
+            2, 4, 4, 2,
+            15,
+            RogueSounds.ROGUE_ARMOR_EQUIP.entry(), () -> { return Ingredient.ofItems(Items.NETHERITE_INGOT); });
+
     public static RegistryEntry<ArmorMaterial> material_warrior_t1 = material(
             "warrior_armor",
             2, 5, 4, 1,
@@ -63,6 +69,12 @@ public class Armors {
             3, 8, 6, 2,
             10,
             RogueSounds.WARRIOR_ARMOR_EQUIP.entry(), () -> { return Ingredient.ofItems(Items.CHAIN); });
+
+    public static RegistryEntry<ArmorMaterial> material_warrior_t3 = material(
+            "netherite_berserker_armor",
+            3, 8, 6, 2,
+            15,
+            RogueSounds.WARRIOR_ARMOR_EQUIP.entry(), () -> { return Ingredient.ofItems(Items.NETHERITE_INGOT); });
 
 
     public static final ArrayList<Armor.Entry> entries = new ArrayList<>();
@@ -79,10 +91,11 @@ public class Armors {
     }
 
 
-    private static final Identifier ATTACK_DAMAGE_ID = Identifier.of("generic.attack_damage");
-    private static final Identifier ATTACK_SPEED_ID = Identifier.of("generic.attack_speed");
-    private static final Identifier KNOCKBACK_ID = Identifier.of("generic.knockback_resistance");
-    private static final Identifier MOVEMENT_SPEED_ID = Identifier.of("generic.movement_speed");
+    private static final Identifier ATTACK_DAMAGE_ID = Identifier.ofVanilla("generic.attack_damage");
+    private static final Identifier ATTACK_SPEED_ID = Identifier.ofVanilla("generic.attack_speed");
+    private static final Identifier KNOCKBACK_ID = Identifier.ofVanilla("generic.knockback_resistance");
+    private static final Identifier MOVEMENT_SPEED_ID = Identifier.ofVanilla("generic.movement_speed");
+    private static final Identifier ARMOR_TOUGHNESS_ID = Identifier.ofVanilla("generic.armor_toughness");
 
     private static ItemConfig.Attribute damageMultiplier(float value) {
         return new ItemConfig.Attribute(
@@ -98,11 +111,11 @@ public class Armors {
                 EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE);
     }
 
-    private static ItemConfig.Attribute knockbackMultiplier(float value) {
+    private static ItemConfig.Attribute knockbackBonus(float value) {
         return new ItemConfig.Attribute(
                 KNOCKBACK_ID.toString(),
                 value,
-                EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE);
+                EntityAttributeModifier.Operation.ADD_VALUE);
     }
 
     private static ItemConfig.Attribute movementSpeed(float value) {
@@ -112,15 +125,33 @@ public class Armors {
                 EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE);
     }
 
-    public static final float rogue_haste_T1 = 0.04F;
-    public static final float rogue_damage_T1 = 0.02F;
-    public static final float rogue_speed_T1 = 0.05F;
-    public static final float rogue_haste_T2 = 0.05F;
-    public static final float rogue_damage_T2 = 0.02F;
-    public static final float rogue_speed_T2 = 0.05F;
-    public static final float warrior_damage_T1 = 0.04F;
-    public static final float warrior_damage_T2 = 0.05F;
-    public static final float knockback_T2 = 0.1F;
+    private static ItemConfig.Attribute toughnessBonus(float value) {
+        return new ItemConfig.Attribute(
+                ARMOR_TOUGHNESS_ID.toString(),
+                value,
+                EntityAttributeModifier.Operation.ADD_VALUE);
+    }
+
+    public static final float rogue_t1_speed = 0.05F;
+    public static final float rogue_t1_haste = 0.04F;
+    // public static final float rogue_t1_damage = 0.02F;
+
+    public static final float rogue_t2_speed = 0.05F;
+    public static final float rogue_t2_haste = 0.05F;
+    public static final float rogue_t2_damage = 0.02F;
+
+    public static final float rogue_t3_speed = 0.05F;
+    public static final float rogue_t3_haste = 0.05F;
+    public static final float rogue_t3_damage = 0.05F;
+
+    public static final float warrior_t1_damage = 0.04F;
+
+    public static final float warrior_t2_damage = 0.05F;
+    public static final float warrior_t2_knockback = 0.1F;
+
+    public static final float warrior_t3_damage = 0.05F;
+    public static final float warrior_t3_toughness = 1F;
+    public static final float warrior_t3_knockback = 0.1F;
 
     public static final Armor.Set RogueArmorSet_t1 = create(
             material_rogue_t1,
@@ -129,17 +160,17 @@ public class Armors {
             RogueArmor::new,
             ItemConfig.ArmorSet.with(
                     new ItemConfig.ArmorSet.Piece(1)
-                            .add(hasteMultiplier(rogue_haste_T1))
-                            .add(movementSpeed(rogue_speed_T1)),
+                            .add(movementSpeed(rogue_t1_speed))
+                            .add(hasteMultiplier(rogue_t1_haste)),
                     new ItemConfig.ArmorSet.Piece(3)
-                            .add(hasteMultiplier(rogue_haste_T1))
-                            .add(movementSpeed(rogue_speed_T1)),
+                            .add(movementSpeed(rogue_t1_speed))
+                            .add(hasteMultiplier(rogue_t1_haste)),
                     new ItemConfig.ArmorSet.Piece(3)
-                            .add(hasteMultiplier(rogue_haste_T1))
-                            .add(movementSpeed(rogue_speed_T1)),
+                            .add(movementSpeed(rogue_t1_speed))
+                            .add(hasteMultiplier(rogue_t1_haste)),
                     new ItemConfig.ArmorSet.Piece(1)
-                            .add(hasteMultiplier(rogue_haste_T1))
-                            .add(movementSpeed(rogue_speed_T1))
+                            .add(movementSpeed(rogue_t1_speed))
+                            .add(hasteMultiplier(rogue_t1_haste))
             ))
             .armorSet();
 
@@ -150,21 +181,46 @@ public class Armors {
             RogueArmor::new,
             ItemConfig.ArmorSet.with(
                     new ItemConfig.ArmorSet.Piece(2)
-                            .add(damageMultiplier(rogue_damage_T2))
-                            .add(hasteMultiplier(rogue_haste_T2))
-                            .add(movementSpeed(rogue_speed_T2)),
+                            .add(movementSpeed(rogue_t2_speed))
+                            .add(hasteMultiplier(rogue_t2_haste))
+                            .add(damageMultiplier(rogue_t2_damage)),
                     new ItemConfig.ArmorSet.Piece(4)
-                            .add(damageMultiplier(rogue_damage_T2))
-                            .add(hasteMultiplier(rogue_haste_T2))
-                            .add(movementSpeed(rogue_speed_T2)),
+                            .add(movementSpeed(rogue_t2_speed))
+                            .add(hasteMultiplier(rogue_t2_haste))
+                            .add(damageMultiplier(rogue_t2_damage)),
                     new ItemConfig.ArmorSet.Piece(4)
-                            .add(damageMultiplier(rogue_damage_T2))
-                            .add(hasteMultiplier(rogue_haste_T2))
-                            .add(movementSpeed(rogue_speed_T2)),
+                            .add(movementSpeed(rogue_t2_speed))
+                            .add(hasteMultiplier(rogue_t2_haste))
+                            .add(damageMultiplier(rogue_t2_damage)),
                     new ItemConfig.ArmorSet.Piece(2)
-                            .add(damageMultiplier(rogue_damage_T2))
-                            .add(hasteMultiplier(rogue_haste_T2))
-                            .add(movementSpeed(rogue_speed_T2))
+                            .add(movementSpeed(rogue_t2_speed))
+                            .add(hasteMultiplier(rogue_t2_haste))
+                            .add(damageMultiplier(rogue_t2_damage))
+            ))
+            .armorSet();
+
+    public static final Armor.Set RogueArmorSet_t3 = create(
+            material_rogue_t3,
+            Identifier.of(RoguesMod.NAMESPACE, "netherite_assassin_armor"),
+            37,
+            RogueArmor::new,
+            ItemConfig.ArmorSet.with(
+                    new ItemConfig.ArmorSet.Piece(2)
+                            .add(movementSpeed(rogue_t3_speed))
+                            .add(hasteMultiplier(rogue_t3_haste))
+                            .add(damageMultiplier(rogue_t3_damage)),
+                    new ItemConfig.ArmorSet.Piece(4)
+                            .add(movementSpeed(rogue_t3_speed))
+                            .add(hasteMultiplier(rogue_t3_haste))
+                            .add(damageMultiplier(rogue_t3_damage)),
+                    new ItemConfig.ArmorSet.Piece(4)
+                            .add(movementSpeed(rogue_t3_speed))
+                            .add(hasteMultiplier(rogue_t3_haste))
+                            .add(damageMultiplier(rogue_t3_damage)),
+                    new ItemConfig.ArmorSet.Piece(2)
+                            .add(movementSpeed(rogue_t3_speed))
+                            .add(hasteMultiplier(rogue_t3_haste))
+                            .add(damageMultiplier(rogue_t3_damage))
             ))
             .armorSet();
 
@@ -175,13 +231,13 @@ public class Armors {
             WarriorArmor::new,
             ItemConfig.ArmorSet.with(
                     new ItemConfig.ArmorSet.Piece(2)
-                            .add(damageMultiplier(warrior_damage_T1)),
+                            .add(damageMultiplier(warrior_t1_damage)),
                     new ItemConfig.ArmorSet.Piece(5)
-                            .add(damageMultiplier(warrior_damage_T1)),
+                            .add(damageMultiplier(warrior_t1_damage)),
                     new ItemConfig.ArmorSet.Piece(4)
-                            .add(damageMultiplier(warrior_damage_T1)),
+                            .add(damageMultiplier(warrior_t1_damage)),
                     new ItemConfig.ArmorSet.Piece(1)
-                            .add(damageMultiplier(warrior_damage_T1))
+                            .add(damageMultiplier(warrior_t1_damage))
             ))
             .armorSet();
 
@@ -192,17 +248,42 @@ public class Armors {
             WarriorArmor::new,
             ItemConfig.ArmorSet.with(
                     new ItemConfig.ArmorSet.Piece(3)
-                            .add(damageMultiplier(warrior_damage_T2))
-                            .add(knockbackMultiplier(knockback_T2)),
+                            .add(damageMultiplier(warrior_t2_damage))
+                            .add(knockbackBonus(warrior_t2_knockback)),
                     new ItemConfig.ArmorSet.Piece(8)
-                            .add(damageMultiplier(warrior_damage_T2))
-                            .add(knockbackMultiplier(knockback_T2)),
+                            .add(damageMultiplier(warrior_t2_damage))
+                            .add(knockbackBonus(warrior_t2_knockback)),
                     new ItemConfig.ArmorSet.Piece(6)
-                            .add(damageMultiplier(warrior_damage_T2))
-                            .add(knockbackMultiplier(knockback_T2)),
+                            .add(damageMultiplier(warrior_t2_damage))
+                            .add(knockbackBonus(warrior_t2_knockback)),
                     new ItemConfig.ArmorSet.Piece(3)
-                            .add(damageMultiplier(warrior_damage_T2))
-                            .add(knockbackMultiplier(knockback_T2))
+                            .add(damageMultiplier(warrior_t2_damage))
+                            .add(knockbackBonus(warrior_t2_knockback))
+            ))
+            .armorSet();
+
+    public static final Armor.Set WarriorArmorSet_t3 = create(
+            material_warrior_t3,
+            Identifier.of(RoguesMod.NAMESPACE, "netherite_berserker_armor"),
+            37,
+            WarriorArmor::new,
+            ItemConfig.ArmorSet.with(
+                    new ItemConfig.ArmorSet.Piece(3)
+                            .add(damageMultiplier(warrior_t3_damage))
+                            .add(toughnessBonus(warrior_t3_toughness))
+                            .add(knockbackBonus(warrior_t3_knockback)),
+                    new ItemConfig.ArmorSet.Piece(8)
+                            .add(damageMultiplier(warrior_t3_damage))
+                            .add(toughnessBonus(warrior_t3_toughness))
+                            .add(knockbackBonus(warrior_t3_knockback)),
+                    new ItemConfig.ArmorSet.Piece(6)
+                            .add(damageMultiplier(warrior_t3_damage))
+                            .add(toughnessBonus(warrior_t3_toughness))
+                            .add(knockbackBonus(warrior_t3_knockback)),
+                    new ItemConfig.ArmorSet.Piece(3)
+                            .add(damageMultiplier(warrior_t3_damage))
+                            .add(toughnessBonus(warrior_t3_toughness))
+                            .add(knockbackBonus(warrior_t3_knockback))
             ))
             .armorSet();